small fix and add readme

+++ KF_tracking/predictNewLocationsOfTracks.m
@@ -12,7 +12,7 @@ function tracks = predictNewLocationsOfTracks(tracks)
 		% Shift the bounding box so that its center is at 
 		% the predicted location.
 		if size(predictedCentroid,2)==2
-			predictedTFCorner = predictedCentroid - single(bbox(3:4)) / 2;
+			predictedTFCorner = predictedCentroid - double(bbox(3:4)) / 2;
 			tracks(i).bbox = [predictedTFCorner, bbox(3:4)];
 		elseif size(predictedCentroid,2)==3
 			predictedTFCorner = predictedCentroid - single(bbox(4:6)) / 2;
